Migrate whitelist table
